requirements.txt
numpy
pandas
tf-keras

opencv-python
mediapipe
deepface

librosa
sounddevice
scipy

openai-whisper

transformers
torch
streamlit
